Add callback to sendCommand/flushPending
This changes the semantics of the `publish` method's callback execution
to strictly mean that the message asked to be published was sent over
the wire (or failed to have been ).

Fixes #51 -- but also significantly changes the semantics (as witnessed
by the required changes to the tests)

As a side note:  there appears to have been an oversight in the test
'Basics should do callback after publish is flushed' where the actual
publish was not being sent, but the callback was being executed by
the following ping.