FIX #309 - Added 'timeout' connection option. The 'timeout' option specifies the number of millis to wait for a connection to occur and for data to be sent by the server. Effectively this means that in worse case the timeout could be 2x what is specified.
Fixed a small issue where if an options argument was provided but was not an object, the client would
try to emit an error, however the emit mechanism was not yet setup, as the client was not yet created.

Also fixed an issue where if the client emitted an error (if it failed to connect to a server, the
client would close the stream, but the client was not closed, and thus preventing the node application
from terminating.