Connect timeout (#315)
FIX #309 - Added timeout connection option. The timeout option specifies the number of millis to wait for a connection to occur and for data to be sent by the server. The client must complete the dial and receive a PONG from the server within this amount of time, or it fails.

Fixed a small issue where if an options argument was provided but was not an object, the client would
try to emit an error, however, the emit mechanism was not yet set up, as the client was not yet created.

Also fixed an issue where if the client emitted an error (if it failed to connect to a server, the client would close the stream, but the client was not closed and thus preventing the node application from terminating.

Also, cleaned up to do what other clients do - currently, other clients await a `PONG` before they
send subs, the node client has queued subs right from the start before a `PING` is sent.