Jitter (#351)
Added the ability to specify jitter parameters. The jitter is an extra amount of time that is random up to specified value which is added to the standard `reconnectTimeWait`. Two values are exposed `reconnectJitter` and `reconnectJitterTLS`. In addition, there's a `reconnectDelayHandler` where a client can specify a value (in milliseconds) that will be used for delay between reconnects.